Identify heart rate, compass, distance offsets
From my initial comparison of the binary data to the PWX file, I've
identified the integer fields which look to be in little endian
format.

The distance field is floating point in the PWX file, but I've found
what seems to be the integer part every 48 bytes. In fact, each sample
seems to end with this 2 byte integer field, while each record seems
to begin with either 01 (gpsstatus?) or the heartrate.